move SessionWorkdir to cmdhelper due to import cycle
